Refactored find_library into build_hunspell. Changed library names used
