+++ .gitignore
@@ -65,6 +65,7 @@ libs/tmp
 
 # Built binary locations
 libs/linux
-libs/darwin
+hunspell/libhunspell-1.7.so.0
+hunspell/libhunspell-1.7.0.dylib
 hunspell/libhunspell-msvc14-x64.lib
 hunspell/libhunspell-msvc14-x86.lib
