+++ .travis.yml
@@ -108,6 +108,9 @@ install:
   - touch .cibuildwheel
   - python setup.py build
   - pip install .
+  # Nuke these to make sure our tests after this point don't use temporary shared libraries
+  - rm -rf build
+  - rm -rf external/build
 
 script:
   - python -m pytest -vv
